refactor: MVC pattern 적용, 구입금액 입력 구현
