--- index.js
@@ -1,4 +1,0 @@
-const App = require('./src/App.js');
-
-const app = new App();
-app.play();
