+++ src/constants.js
@@ -1,7 +1,7 @@
 const INPUT_MESSAGE = {
 	GET_PAYMENT: '구입금액을 입력해 주세요.\n',
-	GET_WINNING: '\n당첨 번호를 입력해 주세요.\n',
-	GET_BONUS: '\n보너스 번호를 입력해 주세요.\n',
+	GET_WINNING_NUMBER: '\n당첨 번호를 입력해 주세요.\n',
+	GET_BONUS_NUMBER: '\n보너스 번호를 입력해 주세요.\n',
 };
 
 const PRINT_MESSAGE = {
