+++ src/model/Game.js
@@ -1,0 +1,9 @@
+class LotteryGame {
+  #payment;
+
+  setPayment(payment) {
+    this.#payment = payment;
+  }
+}
+
+module.exports = LotteryGame;
