+++ src/view/OutputView.js
@@ -1,0 +1,8 @@
+const {
+  PRINT_MESSAGE: {},
+} = require('../constants');
+const { Console } = require('@woowacourse/mission-utils');
+
+const OutputView = {};
+
+module.exports = OutputView;
